first iteration of cache system; taking a little bit too long to load the data; also small refactoring;
