+++ Birt/Birt.App/Pages/Index.cshtml
@@ -10,3 +10,7 @@
         Welcome to the platform built using Wikidata’s SPARQL endpoint. This service allows you to query, compare, visualize, and share large sets of data and knowledge. The platform also provides recommendations for related information in different languages. Through the use of big data techniques, we offer insights and summaries on topics such as the influences of fine arts, dance, and music in specific regions over recent years.
     </p>
 </div>
+
+@section Scripts {
+    <script type="module" src="~/js/main.js" defer></script>
+}
