+++ Birt/Birt.App/Pages/Shared/_Layout.cshtml
@@ -45,7 +45,7 @@
 <script src="https://cdn.jsdelivr.net/npm/rdflib/dist/rdflib.min.js"></script>
 <script src="~/lib/jquery/dist/jquery.min.js"></script>
 <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-<script src="~/js/site.js" asp-append-version="true"></script>
+<script src="~/js/main.js" asp-append-version="true"></script>
 @await RenderSectionAsync("Scripts", required: false)
 </body>
 </html>
