+++ Birt/Birt.App/wwwroot/js/site.js
@@ -1,25 +1,6 @@
 ﻿/*
 * Main display function
 * */
-function getWorksOfArt() {
-    const dropdown = document.getElementById("queryDropdown");
-    const selectedArtist = dropdown.value;
-    if (selectedArtist === "gogh") {
-        getWorksOfArtGogh();
-    } else if (selectedArtist === "vinci") {
-        getWorksOfArtDaVinci(); 
-    } else if (selectedArtist === "similar-gogh") {
-        fetchSimilarArtists('Q5582'); 
-    } else if (selectedArtist === "similar-vinci") {
-        fetchSimilarArtists('Q762'); 
-    } else if (selectedArtist === "influencedByGogh") {
-        fetchArtistsInfluencedByVanGogh();
-    } else if (selectedArtist === "paintingInfluencesbetween1850and1900") {
-        displayPaintingInfluencesBetween1850And1900();
-    } else if (selectedArtist === "paintingsInfluencedByGogh") {
-        displayTopPaintingsInfluencedByGogh();
-    }
-}
 
 /*
 * Works of art by Vincent van Gogh (Q5582)
@@ -74,6 +55,8 @@ async function getWorksOfArtDaVinci() {
 * */
 
 async function displayTopPaintingsInfluencedByGogh() {
+    const loadingIndicator = document.getElementById('loading');
+    loadingIndicator.style.display = 'block'; // Show loading indicator
     const paintersQuery = `
         PREFIX icon: <http://www.iconontology.org/ontology#>
         PREFIX wdt: <http://www.wikidata.org/prop/direct/>
@@ -128,6 +111,8 @@ async function displayTopPaintingsInfluencedByGogh() {
         allPaintingsData.results.bindings.push(...paintingsData.results.bindings);
     }
 
+
+    loadingIndicator.style.display = 'none'; // Hide loading indicator
     displayResults(allPaintingsData);
 }
 /*
@@ -391,32 +376,6 @@ async function displayComparison(data, artist1Name, artist2Name) {
 
     comparisonContainer.appendChild(table);
 }
-function toggleCriteriaForm() {
-    const queryValue = document.getElementById('queryDropdown').value;
-    const criteriaForm = document.getElementById('criteria-form');
-    if (queryValue === 'similar-gogh' || queryValue === 'similar-vinci') {
-        criteriaForm.style.display = 'block';
-    } else {
-        criteriaForm.style.display = 'none';
-    }
-}
-function handleQuery() {
-    const queryValue = document.getElementById('queryDropdown').value;
-    const notableWork = document.getElementById('notable-work').value;
-    console.log(queryValue);
-    if (!queryValue) {
-        alert("Please select a query before searching.");
-        return;
-    }
-    if (queryValue === 'similar-gogh') {
-        fetchSimilarArtists('Q5582');
-        return;
-    }
-    else if (queryValue === 'similar-vinci') {
-        fetchSimilarArtists('Q762');
-        return;
-    }
-}
 
 /*
 * Artists influenced by Van Gogh
@@ -526,162 +485,149 @@ function displayArtistsInfluencedByVanGogh(data) {
 * Painting influences between 1850 and 1900
 * */
 
-async function fetchLabel(wikidataId) {
-    const query = `
-        SELECT ?label WHERE {
-            wd:${wikidataId} rdfs:label ?label.
-            FILTER(LANG(?label) = "en").
-        }
-    `;
-    const url = "https://query.wikidata.org/sparql?query=" + encodeURIComponent(query);
-    const response = await fetch(url, {
-        headers: {'Accept': 'application/sparql-results+json'}
-    });
-    const data = await response.json();
-    return data.results.bindings[0]?.label?.value || wikidataId;
-}
-async function displayPaintingInfluences(data, title = 'Top Painting Influences Between 1850 and 1900') {
-    const loadingIndicator = document.getElementById('loading');
-    loadingIndicator.style.display = 'block'; // Show loading indicator
-    const resultsContainer = document.getElementById('results');
-    resultsContainer.innerHTML = `<h2>${title}</h2>`; // Clear previous results and set title
 
-    const table = document.createElement('table');
-    table.style.width = '100%';
-    table.style.borderCollapse = 'collapse';
-
-    const headerRow = document.createElement('tr');
-    const headers = ['Artist', 'Influence Description', 'Painting'];
-    headers.forEach(headerText => {
-        const header = document.createElement('th');
-        header.textContent = headerText;
-        header.style.border = '1px solid #ddd';
-        header.style.padding = '8px';
-        header.style.textAlign = 'left';
-        header.style.backgroundColor = '#f2f2f2';
-        headerRow.appendChild(header);
-    });
-    table.appendChild(headerRow);
-
-    const addedArtists = new Set();
-
-    for (const item of data.results.bindings) {
-        const artistId = item.artist ? item.artist.value.split('/').pop() : '';
-        const artistLabel = item.artistLabel ? item.artistLabel.value : 'Unknown';
-        const influenceDescriptionId = item.influenceDescription ? item.influenceDescription.value.split('/').pop() : 'N/A';
-        const influenceDescription = influenceDescriptionId !== 'N/A' ? `<a href="https://www.wikidata.org/wiki/${influenceDescriptionId}" target="_blank" rel="noopener noreferrer">${await fetchLabel(influenceDescriptionId)}</a>` : 'N/A';
-
-        if (influenceDescription !== 'N/A' && !addedArtists.has(artistId)) {
-            addedArtists.add(artistId);
-
-            const row = document.createElement('tr');
-
-            const artistCell = document.createElement('td');
-            artistCell.style.border = '1px solid #ddd';
-            artistCell.style.padding = '8px';
-            const artistLink = document.createElement('a');
-            artistLink.href = `https://www.wikidata.org/wiki/${artistId}`;
-            artistLink.target = '_blank';
-            artistLink.rel = 'noopener noreferrer';
-            artistLink.textContent = artistLabel;
-            artistCell.appendChild(artistLink);
-            row.appendChild(artistCell);
-
-            const influenceCell = document.createElement('td');
-            influenceCell.style.border = '1px solid #ddd';
-            influenceCell.style.padding = '8px';
-            influenceCell.innerHTML = influenceDescription;
-            row.appendChild(influenceCell);
-
-            const paintingCell = document.createElement('td');
-            paintingCell.style.border = '1px solid #ddd';
-            paintingCell.style.padding = '8px';
-            if (item.workImage && item.workImage.value) {
-                const paintingLink = document.createElement('a');
-                paintingLink.href = `https://www.wikidata.org/wiki/${item.work.value.split('/').pop()}`;
-                paintingLink.target = '_blank';
-                paintingLink.rel = 'noopener noreferrer';
-                const paintingImg = document.createElement('img');
-                paintingImg.src = item.workImage.value;
-                paintingImg.alt = item.workLabel.value;
-                paintingImg.width = 100;
-                paintingImg.height = 100;
-                paintingLink.appendChild(paintingImg);
-                paintingCell.appendChild(paintingLink);
-            } else {
-                continue; // Skip rows where the painting is not available
-            }
-            row.appendChild(paintingCell);
-
-            table.appendChild(row);
-        }
-    }
-
-    loadingIndicator.style.display = 'none'; // Hide loading indicator
-    resultsContainer.appendChild(table);
-}
-async function displayPaintingInfluencesBetween1850And1900() {
-    const query = `
-        PREFIX icon: <http://www.iconontology.org/ontology#>
-        PREFIX wdt: <http://www.wikidata.org/prop/direct/>
-        PREFIX wd: <http://www.wikidata.org/entity/>
-        PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
-        PREFIX owl: <http://www.w3.org/2002/07/owl#>
-        
-        SELECT DISTINCT ?artist ?artistLabel ?influenceDescription WHERE {
-            ?artist wdt:P106 wd:Q1028181.
-            ?artist wdt:P569 ?birthDate.
-            FILTER(?birthDate >= "1850-01-01"^^xsd:dateTime)
-            FILTER(?birthDate <= "1900-12-31"^^xsd:dateTime)
-            ?influenced wdt:P737 ?artist.
-            OPTIONAL { ?influenced wdt:P1344 ?influenceDescription. }
-            OPTIONAL { ?artist icon:hasInfluence ?influenceDescription. }
-            SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGUAGE],en". }
-        }
-        ORDER BY ?artistLabel
-        LIMIT 250
-    `;
-    const url = "https://query.wikidata.org/sparql?query=" + encodeURIComponent(query);
-    const response = await fetch(url, {
-        headers: {
-            'Accept': 'application/sparql-results+json'
-        }
-    });
-    const data = await response.json();
-
-    let allPaintingsData = { results: { bindings: [] } };
-
-    for (const item of data.results.bindings) {
-        const artistId = item.artist.value.split('/').pop();
-        const paintingsQuery = `
-            SELECT ?work ?workLabel ?workImage WHERE {
-                ?work wdt:P170 wd:${artistId}.  # Created by the artist
-                ?work wdt:P31 wd:Q3305213.  # Instance of painting (Q3305213)
-                OPTIONAL { ?work wdt:P18 ?workImage. }
-                SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGUAGE],en". }
-            }
-            LIMIT 2
-        `;
-        const paintingsUrl = "https://query.wikidata.org/sparql?query=" + encodeURIComponent(paintingsQuery);
-        const paintingsResponse = await fetch(paintingsUrl, {
-            headers: {
-                'Accept': 'application/sparql-results+json'
-            }
-        });
-        const paintingsData = await paintingsResponse.json();
-
-        // Add artistLabel and influenceDescription to each painting
-        paintingsData.results.bindings.forEach(painting => {
-            painting.artistLabel = item.artistLabel;
-            painting.artist = item.artist;
-            painting.influenceDescription = item.influenceDescription;
-        });
-
-        allPaintingsData.results.bindings.push(...paintingsData.results.bindings);
-    }
-
-    displayPaintingInfluences(allPaintingsData, 'Paintings Influences Between 1850 and 1900');
-}
+// async function displayPaintingInfluences(data, title = 'Top Painting Influences Between 1850 and 1900') {
+//     const resultsContainer = document.getElementById('results');
+//     resultsContainer.innerHTML = `<h2>${title}</h2>`; // Clear previous results and set title
+//
+//     const table = document.createElement('table');
+//     table.style.width = '100%';
+//     table.style.borderCollapse = 'collapse';
+//
+//     const headerRow = document.createElement('tr');
+//     const headers = ['Artist', 'Influence Description', 'Painting'];
+//     headers.forEach(headerText => {
+//         const header = document.createElement('th');
+//         header.textContent = headerText;
+//         header.style.border = '1px solid #ddd';
+//         header.style.padding = '8px';
+//         header.style.textAlign = 'left';
+//         header.style.backgroundColor = '#f2f2f2';
+//         headerRow.appendChild(header);
+//     });
+//     table.appendChild(headerRow);
+//
+//     const addedArtists = new Set();
+//
+//     for (const item of data.results.bindings) {
+//         const artistId = item.artist ? item.artist.value.split('/').pop() : '';
+//         const artistLabel = item.artistLabel ? item.artistLabel.value : 'Unknown';
+//         const influenceDescriptionId = item.influenceDescription ? item.influenceDescription.value.split('/').pop() : 'N/A';
+//         const influenceDescription = influenceDescriptionId !== 'N/A' ? `<a href="https://www.wikidata.org/wiki/${influenceDescriptionId}" target="_blank" rel="noopener noreferrer">${await fetchLabel(influenceDescriptionId)}</a>` : 'N/A';
+//
+//         if (influenceDescription !== 'N/A' && !addedArtists.has(artistId)) {
+//             addedArtists.add(artistId);
+//
+//             const row = document.createElement('tr');
+//
+//             const artistCell = document.createElement('td');
+//             artistCell.style.border = '1px solid #ddd';
+//             artistCell.style.padding = '8px';
+//             const artistLink = document.createElement('a');
+//             artistLink.href = `https://www.wikidata.org/wiki/${artistId}`;
+//             artistLink.target = '_blank';
+//             artistLink.rel = 'noopener noreferrer';
+//             artistLink.textContent = artistLabel;
+//             artistCell.appendChild(artistLink);
+//             row.appendChild(artistCell);
+//
+//             const influenceCell = document.createElement('td');
+//             influenceCell.style.border = '1px solid #ddd';
+//             influenceCell.style.padding = '8px';
+//             influenceCell.innerHTML = influenceDescription;
+//             row.appendChild(influenceCell);
+//
+//             const paintingCell = document.createElement('td');
+//             paintingCell.style.border = '1px solid #ddd';
+//             paintingCell.style.padding = '8px';
+//             if (item.workImage && item.workImage.value) {
+//                 const paintingLink = document.createElement('a');
+//                 paintingLink.href = `https://www.wikidata.org/wiki/${item.work.value.split('/').pop()}`;
+//                 paintingLink.target = '_blank';
+//                 paintingLink.rel = 'noopener noreferrer';
+//                 const paintingImg = document.createElement('img');
+//                 paintingImg.src = item.workImage.value;
+//                 paintingImg.alt = item.workLabel.value;
+//                 paintingImg.width = 100;
+//                 paintingImg.height = 100;
+//                 paintingLink.appendChild(paintingImg);
+//                 paintingCell.appendChild(paintingLink);
+//             } else {
+//                 continue; // Skip rows where the painting is not available
+//             }
+//             row.appendChild(paintingCell);
+//
+//             table.appendChild(row);
+//         }
+//     }
+//
+//     loadingIndicator.style.display = 'none'; // Hide loading indicator
+//     resultsContainer.appendChild(table);
+// }
+// async function displayPaintingInfluencesBetween1850And1900() {
+//     const loadingIndicator = document.getElementById('loading');
+//     loadingIndicator.style.display = 'block'; // Show loading indicator
+//     const query = `
+//         PREFIX icon: <http://www.iconontology.org/ontology#>
+//         PREFIX wdt: <http://www.wikidata.org/prop/direct/>
+//         PREFIX wd: <http://www.wikidata.org/entity/>
+//         PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
+//         PREFIX owl: <http://www.w3.org/2002/07/owl#>
+//        
+//         SELECT DISTINCT ?artist ?artistLabel ?influenceDescription WHERE {
+//             ?artist wdt:P106 wd:Q1028181.
+//             ?artist wdt:P569 ?birthDate.
+//             FILTER(?birthDate >= "1850-01-01"^^xsd:dateTime)
+//             FILTER(?birthDate <= "1900-12-31"^^xsd:dateTime)
+//             ?influenced wdt:P737 ?artist.
+//             OPTIONAL { ?influenced wdt:P1344 ?influenceDescription. }
+//             OPTIONAL { ?artist icon:hasInfluence ?influenceDescription. }
+//             SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGUAGE],en". }
+//         }
+//         ORDER BY ?artistLabel
+//         LIMIT 250
+//     `;
+//     const url = "https://query.wikidata.org/sparql?query=" + encodeURIComponent(query);
+//     const response = await fetch(url, {
+//         headers: {
+//             'Accept': 'application/sparql-results+json'
+//         }
+//     });
+//     const data = await response.json();
+//
+//     let allPaintingsData = { results: { bindings: [] } };
+//
+//     for (const item of data.results.bindings) {
+//         const artistId = item.artist.value.split('/').pop();
+//         const paintingsQuery = `
+//             SELECT ?work ?workLabel ?workImage WHERE {
+//                 ?work wdt:P170 wd:${artistId}.  # Created by the artist
+//                 ?work wdt:P31 wd:Q3305213.  # Instance of painting (Q3305213)
+//                 OPTIONAL { ?work wdt:P18 ?workImage. }
+//                 SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGUAGE],en". }
+//             }
+//             LIMIT 2
+//         `;
+//         const paintingsUrl = "https://query.wikidata.org/sparql?query=" + encodeURIComponent(paintingsQuery);
+//         const paintingsResponse = await fetch(paintingsUrl, {
+//             headers: {
+//                 'Accept': 'application/sparql-results+json'
+//             }
+//         });
+//         const paintingsData = await paintingsResponse.json();
+//
+//         // Add artistLabel and influenceDescription to each painting
+//         paintingsData.results.bindings.forEach(painting => {
+//             painting.artistLabel = item.artistLabel;
+//             painting.artist = item.artist;
+//             painting.influenceDescription = item.influenceDescription;
+//         });
+//
+//         allPaintingsData.results.bindings.push(...paintingsData.results.bindings);
+//     }
+//
+//     await displayPaintingInfluences(allPaintingsData, 'Paintings Influences Between 1850 and 1900');
+// }
 async function fetchMessage(property) {
     const response = await fetch('http://localhost:5242/rdf/' + property);
     if (response.ok) {
@@ -693,3 +639,5 @@ async function fetchMessage(property) {
         console.error('Failed to fetch message');
     }
 }
+
+export { getWorksOfArtGogh, getWorksOfArtDaVinci, fetchSimilarArtists, fetchArtistsInfluencedByVanGogh, displayTopPaintingsInfluencedByGogh};
